Test for --docker_extra_volume
